Move `DefaultTrustedSocketFactory`
