+++ app/k9mail/src/main/java/com/fsck/k9/backends/KoinModule.kt
@@ -5,11 +5,9 @@ import app.k9mail.dev.developmentModuleAdditions
 import com.fsck.k9.backend.BackendManager
 import com.fsck.k9.backend.imap.BackendIdleRefreshManager
 import com.fsck.k9.backend.imap.SystemAlarmManager
-import com.fsck.k9.mail.ssl.DefaultTrustedSocketFactory
+import com.fsck.k9.helper.DefaultTrustedSocketFactory
 import com.fsck.k9.mail.store.imap.IdleRefreshManager
 import com.fsck.k9.mail.store.webdav.SniHostSetter
-import javax.net.ssl.SSLSocket
-import javax.net.ssl.SSLSocketFactory
 import org.koin.dsl.module
 
 val backendsModule = module {
